refacto: move tests to their own folders
